requirements/development.txt
-r common.txt
# Debug
debugpy # ???
django-debug-toolbar==4.1.0
# django-cors-headers==4.2.0

# Testings
pytest-django==4.5.2
bpython==0.24

# Documentation
mkdocs==1.4.3
mkdocs-awesome-pages-plugin==2.9.1
mkdocs-material==9.1.19

# For jupiter
django-extensions==3.2.3
ipykernel==6.24.0
ipython==8.8.0
jupyter==1.0.0
notebook==6.5.2
psutil==5.9.4
